right string, but wrong count. why???

+++ 01_ft_printf/ft_putstr.c
@@ -14,7 +14,7 @@
 
 int	ft_putstr(char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
@@ -22,5 +22,5 @@ int	ft_putstr(char *s)
 		write(1, &s[i], 1);
 		i++;
 	}
-	return(i);
+	return (i);
 }
